Add new configuration do add a line break after generic where clause
When a generic `where` clause is broken into multiple lines, it might
be desirable to add a line break after the `where` clause so that the
first requirement isn't inline with the `where` even if line length
isn't exceeded, so that all requirement clauses line up vertically.

## Changes

- Add new `lineBreakAfterGenericWhereClause` configuration that is
applied on `TokenStreamCreator.visit(_ node: GenericWhereClauseSyntax)`.

- Add new unit tests to validate desired behavior.

- Add `XCTAssertDiff` to `PrettyPrintTestCase`.

- Fix some typos.